MM-24634 Style emojis in sidebar with full opacity
