+++ components/sidebar/sidebar_channel/sidebar_channel_link/sidebar_channel_link.tsx
@@ -8,13 +8,15 @@ import classNames from 'classnames';
 
 import {Channel} from 'mattermost-redux/types/channels';
 
+import {mark, trackEvent} from 'actions/diagnostics_actions';
+
 import CopyUrlContextMenu from 'components/copy_url_context_menu';
 import OverlayTrigger from 'components/overlay_trigger';
 
-import {mark, trackEvent} from 'actions/diagnostics_actions';
-import {localizeMessage} from 'utils/utils';
-import {isDesktopApp} from 'utils/user_agent';
 import Constants from 'utils/constants';
+import {wrapEmojis} from 'utils/emoji_utils';
+import {isDesktopApp} from 'utils/user_agent';
+import {localizeMessage} from 'utils/utils';
 
 import ChannelMentionBadge from '../channel_mention_badge';
 import SidebarChannelIcon from '../sidebar_channel_icon';
@@ -138,11 +140,11 @@ export default class SidebarChannelLink extends React.PureComponent<Props, State
     render() {
         const {link, label, channel, unreadMentions, icon, isMuted} = this.props;
 
-        let hoverLabel: JSX.Element = (
+        let labelElement: JSX.Element = (
             <span
                 className={'SidebarChannelLinkLabel'}
             >
-                {label}
+                {wrapEmojis(label)}
             </span>
         );
         if (this.state.showTooltip) {
@@ -151,7 +153,7 @@ export default class SidebarChannelLink extends React.PureComponent<Props, State
                     {label}
                 </Tooltip>
             );
-            hoverLabel = (
+            labelElement = (
                 <OverlayTrigger
                     delayShow={Constants.OVERLAY_TIME_DELAY}
                     placement='top'
@@ -159,7 +161,7 @@ export default class SidebarChannelLink extends React.PureComponent<Props, State
                     onEntering={this.removeTooltipLink}
                 >
                     <div ref={this.gmItemRef}>
-                        {hoverLabel}
+                        {labelElement}
                     </div>
                 </OverlayTrigger>
             );
@@ -175,7 +177,7 @@ export default class SidebarChannelLink extends React.PureComponent<Props, State
                     className={'SidebarChannelLinkLabel_wrapper'}
                     ref={this.labelRef}
                 >
-                    {hoverLabel}
+                    {labelElement}
                 </div>
                 <ChannelMentionBadge
                     channelId={channel.id}
